requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
binaryornot==0.4.4
bleach==6.2.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==45.0.5
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.6
django-allauth==65.12.1
django-browser-reload==1.18.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-jet-reboot==1.3.10
django-markdownx==4.0.9
django-quill-editor==0.1.42
django-redis==6.0.0
django-taggit==6.1.0
django-taggit-autosuggest==0.4.2
django-timezone-field==7.1
django-unfold==0.64.2
django-vite==3.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
Jinja2==3.1.6
kombu==5.5.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
redis==4.6.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
text-unidecode==1.3
Twisted==25.5.0
txaio==25.6.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2
